Sync WriteDumpResWithLen parameter type to size_t, move cast to protocol boundary
WriteDumpResWithLen parameter changed from uint32_t to size_t to
match currentDataLen_ type. This eliminates static_cast<uint32_t>
at both call sites (WriteDumpRes, WriteMainThreadDone).

The only remaining cast is inside WriteDumpResWithLen when filling
DumpResMessage.dataLen (uint32_t, fixed binary protocol), which is
the single protocol boundary conversion point.

DFXLOGI format specifier updated from %{public}u to %{public}zu
to match size_t parameter.

No unittest impact: WriteDumpResWithLen is private (no test calls
it directly), WriteDumpRes/WriteMainThreadDone public signatures
are unchanged, resMsg.dataLen remains uint32_t in DumpResMessage.

+++ tools/process_dump/dfx_buffer_writer.h
@@ -55,7 +55,7 @@ public:
     void RemoveFileIfNeed(const std::string& dirPath) const;
     int GetBufFd() const;
 private:
-    bool WriteDumpResWithLen(int32_t dumpRes, uint32_t dataLen);
+    bool WriteDumpResWithLen(int32_t dumpRes, size_t dataLen);
     void WriteToBuffer(const std::string& msg);
 
     DfxBufferWriter() = default;
